+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.1.0 - 2022-07-02
+
+### Added
+- Baseline configurations for VPC using Python/Boto3
+- Lint check for Python files
+- Internet/NAT gateway for Boto3 baseline
+- EIP for Boto3 baseline
+- Second NAT Gateway for Terraform baseline
+
 ## 0.0.2 - 2022-06-27
 
 ### Fixed